requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
feedparser==6.0.11
psycopg2-binary==2.9.9
google-generativeai==0.8.5
schedule==1.2.2
python-dotenv==1.0.1
